move local_network into testutils
