+++ darkside-tests/tests/tests.rs
@@ -17,10 +17,10 @@ use zingolib::get_base_address_macro;
 // use zingolib::testutils::chain_generics::conduct_chain::ConductChain as _;
 // use zingolib::testutils::chain_generics::with_assertions::to_clients_proposal;
 use zingolib::testutils::lightclient::from_inputs;
+use zingolib::testutils::local_network::ZingolibLocalNetwork;
 use zingolib::testutils::scenarios::ClientBuilder;
 use zingolib::testutils::scenarios::LIGHTWALLETD_BIN;
 use zingolib::wallet::balance::AccountBalance;
-use zingolib::wallet::local_network::ZingolibLocalNetwork;
 
 #[ignore = "darkside bug, invalid block hash length in tree states"]
 #[tokio::test]
