+++ zingo-cli/src/regtest.rs
@@ -2,9 +2,9 @@
 //! This module contains all regtest-specific functionality
 
 use std::path::PathBuf;
+use zingolib::testutils::local_network::ZingolibLocalNetwork;
 use zingolib::testutils::scenarios::{LIGHTWALLETD_BIN, ZCASH_CLI_BIN, ZCASHD_BIN};
 use zingolib::testvectors::REG_O_ADDR_FROM_ABANDONART;
-use zingolib::wallet::local_network::ZingolibLocalNetwork;
 use zingolib::zingo_infra_services::LocalNet;
 use zingolib::zingo_infra_services::indexer::{Lightwalletd, LightwalletdConfig};
 use zingolib::zingo_infra_services::validator::{Zcashd, ZcashdConfig};
