+++ zingolib/src/wallet/transaction.rs
@@ -10,9 +10,10 @@ use pepper_sync::wallet::{
 use super::LightWallet;
 use super::error::{FeeError, RemovalError, SpendError};
 use super::summary::data::{SendType, TransactionKind};
+#[cfg(any(test, feature = "test-elevation"))]
+use crate::config::ZENNIES_FOR_ZINGO_REGTEST_ADDRESS;
 use crate::config::{
-    ChainType, ZENNIES_FOR_ZINGO_DONATION_ADDRESS, ZENNIES_FOR_ZINGO_REGTEST_ADDRESS,
-    ZENNIES_FOR_ZINGO_TESTNET_ADDRESS,
+    ChainType, ZENNIES_FOR_ZINGO_DONATION_ADDRESS, ZENNIES_FOR_ZINGO_TESTNET_ADDRESS,
 };
 
 impl LightWallet {
@@ -126,6 +127,7 @@ impl LightWallet {
         let zfz_address = match self.network {
             ChainType::Mainnet => ZENNIES_FOR_ZINGO_DONATION_ADDRESS,
             ChainType::Testnet => ZENNIES_FOR_ZINGO_TESTNET_ADDRESS,
+            #[cfg(any(test, feature = "test-elevation"))]
             ChainType::Regtest(_) => ZENNIES_FOR_ZINGO_REGTEST_ADDRESS,
         };
